requirements.txt
streamlit
fpdf2
pandas